feat(#66): 스터디 그룹 멤버 주간 공부기록 API 작업 완료 - Datasource, Repository, Module, UseCase 등 추가 및 타입 수정

+++ app/src/main/java/com/ssafy/neegongnaegong/domain/model/studygroup/StudyMemberInfo.kt
@@ -1,6 +1,6 @@
 package com.ssafy.neegongnaegong.domain.model.studygroup
 
 data class StudyMemberInfo(
-    val studyGroupId: Int,
-    val targetUserId: Int
+    val studyGroupId: Long,
+    val targetUserId: Long
 )
